Rename src to neural_network
